backend_python/requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pandas==2.2.0
openpyxl==3.1.2
fpdf==1.7.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
